+++ classes/local/last_merge.php
@@ -120,4 +120,63 @@ class last_merge {
     public function tome(): null|stdClass {
         return $this->tome;
     }
+
+    /**
+     * Lists deletable users according to this plugin overview.
+     *
+     * Deletable users match these conditions:
+     *
+     * 1. Are not deleted yet.
+     * 2. Were merged as user to remove.
+     * 3. Are suspended.
+     * 4. In presence of a merge as user to keep, the merge as user to remove was later.
+     * 5. Both merges were successful.
+     *
+     * @throws dml_exception
+     */
+    public static function list_all_deletable_users(): array {
+        global $DB;
+        $sql = "
+         SELECT mremove2.fromuserid,
+                mremove2.timemodified AS fromtimemodified,
+                mkeep2.touserid,
+                mkeep2.timemodified AS totimemodified
+           FROM (SELECT mremove.fromuserid,
+                        mremove.timemodified
+                   FROM (SELECT m.fromuserid,
+                                m.timemodified,
+                                m.success,
+                                ROW_NUMBER() OVER (PARTITION BY m.fromuserid ORDER BY m.timemodified DESC) AS rownumber
+                           FROM {tool_mergeusers} m
+                        ) mremove
+                   JOIN {user} u ON (u.id = mremove.fromuserid)
+                  WHERE mremove.rownumber = :rownumber1
+                    AND u.deleted = :nodeleted
+                    AND u.suspended = :suspended 
+                    AND mremove.success = :success1
+                  ) mremove2
+            LEFT JOIN (SELECT
+                           mkeep.touserid,
+                           mkeep.timemodified
+                       FROM (SELECT
+                                 m.touserid,
+                                 m.timemodified,
+                                 m.success,
+                                 ROW_NUMBER() OVER (PARTITION BY m.touserid ORDER BY m.timemodified DESC) AS rownumber
+                             FROM {tool_mergeusers} m) mkeep
+                       WHERE mkeep.rownumber = :rownumber2
+                         AND mkeep.success = :success2
+            ) mkeep2 ON (mremove2.fromuserid = mkeep2.touserid AND mkeep2.timemodified < mremove2.timemodified)";
+
+        $params = [
+            'rownumber1' => 1,
+            'rownumber2' => 1,
+            'nodeleted' => 0,
+            'suspended' => 1,
+            'success1' => 1,
+            'success2' => 1,
+        ];
+
+        return $DB->get_records_sql($sql, $params);
+    }
 }